Fix up import issues
